openssl: fix compiler warning
tv_sec and tv_usec are declared as long on Windows. This is a 32-bit
type, which means that it disagrees with size_t on 64-bit Windows.

Let's just cast the right-hand side to long and shut up the compiler.

+++ lib/vtls/openssl.c
@@ -293,11 +293,13 @@ static CURLcode Curl_ossl_seed(struct Curl_easy *data)
     for(i = 0, i_max = len / sizeof(struct timeval); i < i_max; ++i) {
       struct timeval tv = curlx_tvnow();
       Curl_wait_ms(1);
-      tv.tv_sec *= i + 1;
-      tv.tv_usec *= i + 2;
-      tv.tv_sec ^= ((curlx_tvnow().tv_sec + curlx_tvnow().tv_usec) *
+      tv.tv_sec *= (long)i + 1;
+      tv.tv_usec *= (long)i + 2;
+      tv.tv_sec ^= (long)
+                   ((curlx_tvnow().tv_sec + curlx_tvnow().tv_usec) *
                     (i + 3)) << 8;
-      tv.tv_usec ^= ((curlx_tvnow().tv_sec + curlx_tvnow().tv_usec) *
+      tv.tv_usec ^= (long)
+                    ((curlx_tvnow().tv_sec + curlx_tvnow().tv_usec) *
                      (i + 4)) << 16;
       memcpy(&randb[i * sizeof(struct timeval)], &tv, sizeof(struct timeval));
     }
